Add updates from review
