+++ src/utils.ts
@@ -11,8 +11,9 @@ export { default as hasAriaLabel } from "./utils/hasAriaLabel";
 export { default as hasContent } from "./utils/hasContent";
 export { default as hasOnDirective } from "./utils/hasOnDirective";
 export { default as hasOnDirectives } from "./utils/hasOnDirectives";
-export { default as interactiveHandlers } from "./utils/interactiveHandlers";
+export { default as interactiveHandlers } from "./utils/interactiveHandlers.json";
 export { default as isAriaHidden } from "./utils/isAriaHidden";
+export { default as isCustomComponent } from "./utils/isCustomComponent";
 export { default as isAttribute } from "./utils/isAttribute";
 export { default as isHiddenFromScreenReader } from "./utils/isHiddenFromScreenReader";
 export { default as isInteractiveElement } from "./utils/isInteractiveElement";
