+++ src/utils/interactiveHandlers.json
@@ -1,4 +1,4 @@
-export default [
+[
   "click",
   "contextmenu",
   "dblclick",
